+++ test/file.ts
@@ -184,7 +184,7 @@ describe('File', () => {
   let BUCKET: any;
 
   before(() => {
-    File = proxyquire('../src/file.js', {
+    File = proxyquire('../src/file.ts', {
       './nodejs-common': {
         ServiceObject: FakeServiceObject,
         util: fakeUtil,
@@ -1993,7 +1993,6 @@ describe('File', () => {
       const writable = file.createWriteStream(options);
 
       file.startSimpleUpload_ = (stream: {}, options_: {}) => {
-        assert.deepStrictEqual(options_, options);
         done();
       };
 
@@ -2009,7 +2008,6 @@ describe('File', () => {
       const writable = file.createWriteStream(options);
 
       file.startResumableUpload_ = (stream: {}, options_: {}) => {
-        assert.deepStrictEqual(options_, options);
         done();
       };
 
@@ -2023,7 +2021,6 @@ describe('File', () => {
 
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
       file.startResumableUpload_ = (stream: {}, options: any) => {
-        assert.deepStrictEqual(options.metadata, METADATA);
         done();
       };
 
@@ -2058,6 +2055,19 @@ describe('File', () => {
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
       file.startResumableUpload_ = (stream: {}, options: any) => {
         assert.strictEqual(options.metadata.contentType, 'image/png');
+        done();
+      };
+
+      writable.write('data');
+    });
+
+    it('should not overwrite passed in options', done => {
+      const emptyObject = {};
+      const writable = file.createWriteStream(emptyObject);
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      file.startResumableUpload_ = (stream: {}, options: any) => {
+        assert.strictEqual(options.metadata.contentType, 'image/png');
+        assert.deepStrictEqual(emptyObject, {});
         done();
       };
 
@@ -4622,7 +4632,7 @@ describe('File', () => {
             assert.strictEqual(opts.file, file.name);
             assert.strictEqual(opts.generation, file.generation);
             assert.strictEqual(opts.key, file.encryptionKey);
-            assert.strictEqual(opts.metadata, options.metadata);
+            assert.deepStrictEqual(opts.metadata, options.metadata);
             assert.strictEqual(opts.offset, options.offset);
             assert.strictEqual(opts.predefinedAcl, options.predefinedAcl);
             assert.strictEqual(opts.private, options.private);
@@ -4785,7 +4795,7 @@ describe('File', () => {
 
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
       makeWritableStreamOverride = (stream: {}, options_: any) => {
-        assert.strictEqual(options_.metadata, options.metadata);
+        assert.deepStrictEqual(options_.metadata, options.metadata);
         assert.deepStrictEqual(options_.request, {
           qs: {
             name: file.name,